+++ src/components/Header/Header.tsx
@@ -29,8 +29,20 @@ import {
   refreshAccessToken,
   tokenRefreshLink,
 } from "./utils/headerUtils";
+import gaUtils from "./utils/googleAnalyticsUtils";
 
-export const Header = chakra(() => {
+export interface GAOptionProps {
+  debug?: boolean;
+  standardImplementation?: boolean;
+  testMode?: boolean;
+  titleCase?: boolean;
+}
+export interface HeaderProps {
+  gaOptions?: GAOptionProps;
+  isProduction?: boolean;
+}
+
+export const Header = chakra(({ gaOptions = {}, isProduction = true }) => {
   const [patronName, setPatronName] = useState<string>("");
   const { isLargerThanMobile, isLargerThanLarge } = useNYPLBreakpoints();
   const styles = useMultiStyleConfig("Header", {});
@@ -50,6 +62,23 @@ export const Header = chakra(() => {
       setPatronName(fullName);
     }
   };
+
+  useEffect(() => {
+    if (!(window as any)?.ga) {
+      // @TODO not sure if we still want this to be logged.
+      // console.log('Analytics not available - loading through React.');
+      console.info(
+        "NYPL Reservoir Header: Loading Google Analytics through the Header component."
+      );
+      const gaOpts = {
+        testMode: !isProduction,
+        ...gaOptions,
+      };
+
+      // Passing false to get the dev GA code.
+      gaUtils.initialize(gaOpts, isProduction);
+    }
+  }, [gaOptions, isProduction]);
 
   useEffect(() => {
     // After mounting,look for a cookie named "nyplIdentityPatron"
@@ -73,6 +102,7 @@ export const Header = chakra(() => {
             <Link
               aria-label="The New York Public Library"
               href="https://nypl.org"
+              onClick={() => gaUtils.trackEvent("Click Logo", "")}
               __css={styles.logo}
             >
               <Logo