+++ src/components/Header/components/HeaderLowerNav.tsx
@@ -1,43 +1,30 @@
-import React from "react";
 import { Box, chakra, useStyleConfig } from "@chakra-ui/react";
+import React from "react";
 
 import Link from "../../Link/Link";
 import List from "../../List/List";
 import HeaderSearchButton from "./HeaderSearchButton";
-import { lowerNavLinks } from "../utils/headerUtils";
+import { siteNavLinks } from "../utils/headerUtils";
+import gaUtils from "../utils/googleAnalyticsUtils";
 
 const HeaderLowerNav = chakra(() => {
   const styles = useStyleConfig("HeaderLowerNav");
+  const listItems = siteNavLinks.map(({ href, text }) => (
+    <Link
+      href={href}
+      key={text}
+      onClick={() => gaUtils.trackEvent("Go to...", text)}
+    >
+      {text}
+    </Link>
+  ));
 
   return (
     <Box as="nav" aria-label="Header bottom links" __css={styles}>
       <List
         id="header-nav-lower"
         inline
-        listItems={[
-          <Link href={lowerNavLinks.books} key="booksMusicMoviesLink">
-            Books/Music/Movies
-          </Link>,
-          <Link href={lowerNavLinks.research} key="researchLink">
-            Research
-          </Link>,
-          <Link href={lowerNavLinks.education} key="educationLink">
-            Education
-          </Link>,
-          <Link href={lowerNavLinks.events} key="eventsLink">
-            Events
-          </Link>,
-          <Link href={lowerNavLinks.connect} key="connectLink">
-            Connect
-          </Link>,
-          <Link href={lowerNavLinks.give} key="giveLink">
-            Give
-          </Link>,
-          <Link href={lowerNavLinks.getHelp} key="getHelpLink">
-            Get Help
-          </Link>,
-          <HeaderSearchButton key="search" />,
-        ]}
+        listItems={[...listItems, <HeaderSearchButton key="search" />]}
         noStyling
         type="ul"
       />