+++ src/components/Header/components/HeaderMobileIconNav.tsx
@@ -1,11 +1,12 @@
 import React from "react";
 import { chakra, Flex, useStyleConfig } from "@chakra-ui/react";
 
-import Icon from "../../Icons/Icon";
-import Link from "../../Link/Link";
 import HeaderLoginButton from "./HeaderLoginButton";
 import HeaderMobileNavButton from "./HeaderMobileNavButton";
 import HeaderSearchButton from "./HeaderSearchButton";
+import Icon from "../../Icons/Icon";
+import Link from "../../Link/Link";
+import gaUtils from "../utils/googleAnalyticsUtils";
 
 const HeaderMobileIconNav = chakra(() => {
   const styles = useStyleConfig("HeaderMobileIconNav");
@@ -16,6 +17,7 @@ const HeaderMobileIconNav = chakra(() => {
       <Link
         aria-label="NYPL Locations Near Me"
         href="https://nypl.org/locations"
+        onClick={() => gaUtils.trackEvent("Click", "Mobile Locations Button")}
       >
         <Icon name="locator" size="large" title="NYPL Locator" />
       </Link>