+++ src/components/Header/components/HeaderUpperNav.tsx
@@ -1,10 +1,11 @@
-import React from "react";
 import { Box, chakra, useMultiStyleConfig } from "@chakra-ui/react";
+import React from "react";
 
+import HeaderLoginButton from "./HeaderLoginButton";
 import Icon from "../../Icons/Icon";
 import Link from "../../Link/Link";
 import List from "../../List/List";
-import HeaderLoginButton from "./HeaderLoginButton";
+import gaUtils from "../utils/googleAnalyticsUtils";
 import { upperNavLinks } from "../utils/headerUtils";
 
 const HeaderUpperNav = chakra(() => {
@@ -17,32 +18,52 @@ const HeaderUpperNav = chakra(() => {
         inline
         listItems={[
           <HeaderLoginButton key="login" />,
-          <Link href={upperNavLinks.locations} key="locationsLink">
-            Locations
-          </Link>,
-          <Link href={upperNavLinks.libraryCard} key="libraryCardLink">
-            Get A Library Card
+          <Link
+            href={upperNavLinks.locations.href}
+            key="locationsLink"
+            onClick={() => gaUtils.trackEvent("Locations", "Header Top Links")}
+          >
+            {upperNavLinks.locations.text}
           </Link>,
           <Link
-            href={upperNavLinks.emailUpdates}
+            href={upperNavLinks.libraryCard.href}
+            key="libraryCardLink"
+            onClick={() =>
+              gaUtils.trackEvent("Get a Library Card", "Header Top Links")
+            }
+          >
+            {upperNavLinks.libraryCard.text}
+          </Link>,
+          <Link
+            href={upperNavLinks.emailUpdates.href}
             key="emailUpdatesLink"
+            onClick={() =>
+              // This is a new event not in the original NYPL Header.
+              // The old implementation used a dropdown for subcription.
+              gaUtils.trackEvent("Subscribe", "Subscription")
+            }
             __css={styles.emailUpdatesLink}
           >
             <>
-              Get Email Updates
+              {upperNavLinks.emailUpdates.text}
               <Icon name="arrow" align="right" size="small" />
             </>
           </Link>,
           <Link
-            href={upperNavLinks.donate}
+            href={upperNavLinks.donate.href}
             key="donateLink"
+            onClick={() => gaUtils.trackEvent("Donate", "Header Top Links")}
             type="button"
             __css={styles.donateLink}
           >
-            Donate
+            {upperNavLinks.donate.text}
           </Link>,
-          <Link href={upperNavLinks.shop} key="shopLink">
-            Shop
+          <Link
+            href={upperNavLinks.shop.href}
+            key="shopLink"
+            onClick={() => gaUtils.trackEvent("Shop", "Header Top Links")}
+          >
+            {upperNavLinks.shop.text}
           </Link>,
         ]}
         noStyling