+++ src/components/Header/utils/headerUtils.ts
@@ -16,23 +16,57 @@ export const patronApiUrl =
   "https://platform.nypl.org/api/v0.1/auth/patron/tokens/";
 export const tokenRefreshLink = `${authServerDomain}/refresh`;
 export const upperNavLinks = {
-  locations: "https://www.nypl.org/locations",
-  libraryCard: "https://www.nypl.org/library-card/new",
-  emailUpdates:
-    "https://pages.email.nypl.org/page.aspx?QS=3935619f7de112ef7250fe02b84fb2f9ab74e4ea015814b7",
-  donate:
-    "https://secure.nypl.org/site/Donation2?7825.donation=form1&df_id=7825&mfc_pref=T&s_src=FRQ18ZZ_TNN",
-  shop: "https://shop.nypl.org/?utm_campaign=NYPLHeaderButton&utm_source=nypl.org&utm_medium=referral",
+  locations: {
+    href: "https://www.nypl.org/locations",
+    text: "Locations",
+  },
+  libraryCard: {
+    href: "https://www.nypl.org/library-card/new",
+    text: "Get A Library Card",
+  },
+  emailUpdates: {
+    href: "https://pages.email.nypl.org/page.aspx?QS=3935619f7de112ef7250fe02b84fb2f9ab74e4ea015814b7",
+    text: "Get Email Updates",
+  },
+  donate: {
+    href: "https://secure.nypl.org/site/Donation2?7825.donation=form1&df_id=7825&mfc_pref=T&s_src=FRQ18ZZ_TNN",
+    text: "Donate",
+  },
+  shop: {
+    href: "https://shop.nypl.org/?utm_campaign=NYPLHeaderButton&utm_source=nypl.org&utm_medium=referral",
+    text: "Shop",
+  },
 };
-export const lowerNavLinks = {
-  books: "https://www.nypl.org/books-music-movies",
-  research: "https://www.nypl.org/research",
-  education: "https://www.nypl.org/education",
-  events: "https://www.nypl.org/events",
-  connect: "https://www.nypl.org/connect",
-  give: "https://www.nypl.org/give",
-  getHelp: "https://www.nypl.org/get-help",
-};
+export const siteNavLinks = [
+  {
+    href: "https://www.nypl.org/books-music-movies",
+    text: "Books/Music/Movies",
+  },
+  {
+    href: "https://www.nypl.org/research",
+    text: "Research",
+  },
+  {
+    href: "https://www.nypl.org/education",
+    text: "Education",
+  },
+  {
+    href: "https://www.nypl.org/events",
+    text: "Events",
+  },
+  {
+    href: "https://www.nypl.org/connect",
+    text: "Connect",
+  },
+  {
+    href: "https://www.nypl.org/give",
+    text: "Give",
+  },
+  {
+    href: "https://www.nypl.org/get-help",
+    text: "Get Help",
+  },
+];
 export const loginLinks = {
   catalog: "https://browse.nypl.org/iii/encore/myaccount",
   research: "https://catalog.nypl.org/patroninfo/top",